+++ CarTrackingLoop.py
@@ -8,34 +8,21 @@ from ultralytics.solutions.solutions import SolutionAnnotator
 import logging
 
 from Clients.sql import SqlClient
-from EventManager import EventManager, NullEventManager, SqlEventManager
-from TrackingManager import TrackingManager
+from tracking_core.annotation_util import render_boxes, render_text, render_traffic_light, render_zones
+from tracking_core.draw_util import inverse_text
+from tracking_core.EventManager import EventManager, NullEventManager, SqlEventManager
+from tracking_core.TrackingManager import TrackingManager
 from constants import (
-    BLACK_LIGHT,
-    BOX_LINE_THICKNESS,
-    BOX_TEXT_COLOR,
-    BOX_TEXT_SCALE,
-    BOX_TEXT_THICKNESS,
     COUNT_ZONE,
     FRAME_BUFFER_FILE,
-    GREEN_LIGHT,
-    LINE_COLOR,
-    LINE_THICKNESS,
     MODEL_NAME,
     MOUSE_COLOR,
     OUTPUT_VIDEO,
-    RED_LIGHT,
     TEXT_COLOR,
-    TEXT_LINE_HEIGHT,
-    TEXT_SCALE,
-    TEXT_THICKNESS,
     TRACKING_CLASSES,
-    TRACKING_LABELS,
     WINDOW_NAME,
-    YELLOW_LIGHT,
-    ZONE_CLEAR_COUNTDOWN_SEC,
 )
-from structures import LightColor, TrackingData, TrackingFrame
+from tracking_core.structures import TrackingFrame
 
 logging.getLogger("ultralytics").setLevel(logging.ERROR)
 
@@ -93,62 +80,6 @@ class Experiment:
         self.sql = None
         self.paused = False
 
-    def _interpolate_color(
-        self, start_color, end_color, t: float
-    ) -> tuple[int, int, int]:
-        """Linearly interpolate between two BGR colors"""
-        return tuple(
-            [
-                int(start + (end - start) * t)
-                for start, end in zip(start_color, end_color)
-            ]
-        )
-
-    def _render_zones(self, frame: TrackingFrame, frame_image: np.ndarray):
-
-        line_color = LINE_COLOR
-
-        if frame.zone_clear_time > 0:
-            elapsed = (frame.time_offset - frame.zone_clear_time) / 1000
-            t = min(elapsed / ZONE_CLEAR_COUNTDOWN_SEC, 1.0)
-            line_color = self._interpolate_color(LINE_COLOR, (0, 0, 255), t)
-        elif frame.zone_clear_time == -1:
-            # Flash every 500ms
-            if frame.time_offset % 1000 < 500:
-                line_color = (0, 0, 255)
-            else:
-                return
-
-        cv2.polylines(
-            frame_image,
-            [np.array(COUNT_ZONE, dtype=np.int32)],
-            isClosed=True,
-            color=line_color,
-            thickness=LINE_THICKNESS,
-        )
-
-    def _render_text(self, frame: TrackingFrame, frame_image: np.ndarray):
-        current_y = 50
-
-        def out(text):
-            nonlocal current_y
-            cv2.putText(
-                frame_image,
-                text,
-                (10, current_y),
-                cv2.FONT_HERSHEY_SIMPLEX,
-                TEXT_SCALE,
-                TEXT_COLOR,
-                TEXT_THICKNESS,
-            )
-            current_y += TEXT_LINE_HEIGHT
-
-        out(f"Frame: {frame.frame_index}")
-        out(f"Time: {(frame.time_offset/1000):.2f}s")
-        out(f"Objects in: {frame.in_count}")
-        out(f"Objects out: {frame.passed_through_count}")
-        out(f"Think time: {frame.frame_processing_time_ms}ms")
-
     def _render_help(self, frame_image: np.ndarray):
         current_y = 500
 
@@ -178,145 +109,6 @@ class Experiment:
         out("Forward in time (->)", self.paused)
         out("Quit (Q)")
 
-    def _center_text(
-        self, frame_image: np.ndarray, text, xy_center, color, thickness=1, scale=1.0
-    ):
-        (text_width, text_height), _ = cv2.getTextSize(
-            text, cv2.FONT_HERSHEY_SIMPLEX, scale, thickness
-        )
-
-        cv2.putText(
-            frame_image,
-            text,
-            (xy_center[0] - text_width // 2, xy_center[1] + text_height // 2),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            scale,
-            color,
-            thickness,
-            cv2.LINE_AA,
-        )
-
-    def _inverse_text(
-        self, frame_image: np.ndarray, text, xy, color, thickness=1, scale=1.0
-    ):
-        (text_width, text_height), baseline = cv2.getTextSize(
-            text, cv2.FONT_HERSHEY_SIMPLEX, scale, thickness
-        )
-        baseline += 2
-        box_coords1 = (xy[0] - 4, xy[1] - text_height - 4)
-        box_coords2 = (xy[0] + text_width + 4, xy[1] + baseline)
-
-        cv2.rectangle(frame_image, box_coords1, box_coords2, color, cv2.FILLED)
-        cv2.putText(
-            frame_image,
-            text,
-            xy,
-            cv2.FONT_HERSHEY_SIMPLEX,
-            scale,
-            (0, 0, 0),
-            thickness,
-            cv2.LINE_AA,
-        )
-
-    def _render_traffic_light(self, frame: TrackingFrame, frame_image: np.ndarray):
-        x, y, w, h = 1800, 1, 80, 300  # x, y coordinates, width, height
-        padding = 7
-        circle_radius = (w - 2 * padding) // 2
-
-        cv2.rectangle(frame_image, (x, y), (x + w, y + h), (0, 0, 0), -1)
-
-        cv2.circle(
-            frame_image,
-            (x + w // 2, y + h // 4),
-            circle_radius,
-            RED_LIGHT if frame.light_color == LightColor.RED else BLACK_LIGHT,
-            -1,
-        )
-
-        cv2.circle(
-            frame_image,
-            (x + w // 2, y + h // 2),
-            circle_radius,
-            YELLOW_LIGHT if frame.light_color == LightColor.YELLOW else BLACK_LIGHT,
-            -1,
-        )
-        cv2.circle(
-            frame_image,
-            (x + w // 2, y + 3 * h // 4),
-            circle_radius,
-            GREEN_LIGHT if frame.light_color == LightColor.GREEN else BLACK_LIGHT,
-            -1,
-        )
-
-        if (
-            frame.light_change_time
-            and frame.light_duration
-            and (
-                frame.light_color == LightColor.RED
-                or frame.light_color == LightColor.GREEN
-            )
-        ):
-
-            elapsed = round((frame.time_offset - frame.light_change_time) / 1000)
-            remaining = max(0, round(frame.light_duration / 1000) - elapsed)
-
-            if frame.light_color == LightColor.RED:
-                red_text = f"{elapsed}s"
-                green_text = f"{remaining}s"
-                red_color = (255, 255, 255)
-                green_color = (0, 0, 0)
-            else:
-                green_text = f"{elapsed}s"
-                red_text = f"{remaining}s"
-                red_color = (0, 0, 0)
-                green_color = (255, 255, 255)
-
-            self._center_text(
-                frame_image,
-                red_text,
-                (x + w // 2, y + h // 4),
-                color=green_color,
-                scale=0.6,
-            )
-            self._center_text(
-                frame_image,
-                green_text,
-                (x + w // 2, y + 3 * h // 4),
-                color=red_color,
-                scale=0.6,
-            )
-
-    def _render_single_box(self, frame_image: np.ndarray, td: TrackingData, color):
-        x1, y1, x2, y2 = td.box
-        label = TRACKING_LABELS[td.class_id]
-        label = f"{label}/{td.id}"
-
-        cv2.rectangle(
-            frame_image,
-            (x1, y1),
-            (x2, y2),
-            color,
-            BOX_LINE_THICKNESS,
-            cv2.LINE_AA,
-        )
-        self._inverse_text(
-            frame_image,
-            label,
-            (x1, y1),
-            BOX_TEXT_COLOR,
-            BOX_TEXT_THICKNESS,
-            BOX_TEXT_SCALE,
-        )
-
-    def _render_boxes(self, frame: TrackingFrame, frame_image: np.ndarray):
-
-        for td in frame.tracking_data:
-            if not td.in_zone:
-                # Render only objects that are not in the counting zone
-                continue
-
-            self._render_single_box(frame_image, td, (0, 0, 255))
-
     def _render_mouse(self, frame_image: np.ndarray):
         if not hasattr(self, "mouse_pos"):
             return
@@ -335,7 +127,7 @@ class Experiment:
             2,
         )
 
-        self._inverse_text(
+        inverse_text(
             frame_image,
             f"{self.mouse_pos[0]},{self.mouse_pos[1]}",
             (self.mouse_pos[0] + 10, self.mouse_pos[1] - 10),
@@ -346,13 +138,13 @@ class Experiment:
 
     def _render_frame(self, frame: TrackingFrame, frame_image: np.ndarray):
         if self.show_zones:
-            self._render_zones(frame, frame_image)
+            render_zones(frame, frame_image)
         if self.show_light:
-            self._render_traffic_light(frame, frame_image)
+            render_traffic_light(frame, frame_image)
         if self.show_text:
-            self._render_text(frame, frame_image)
+            render_text(frame, frame_image)
         if self.show_boxes:
-            self._render_boxes(frame, frame_image)
+            render_boxes(frame, frame_image)
         if self.show_mouse:
             self._render_mouse(frame_image)
