+++ webapp/backend/requirements.txt
@@ -1,2 +1,9 @@
 fastapi
 uvicorn
+
+opencv-python
+flet
+ultralytics
+cvzone
+flet[all]
+psycopg2-binary
