--- webapp/backend/start
@@ -1,1 +1,0 @@
-uvicorn main:app --reload
