--- webapp/frontend/start
@@ -1,1 +1,0 @@
-npm run dev
